調整corner拖拉範圍

+++ js/iphone4_layout_1.js
@@ -10,8 +10,8 @@ for (var i=1; i<=23; i++) {
 	$("#mCanvas").after($destDiv);    
 	$destDiv.corner("999px");
 	// Adjust position according to mCanvas
-	_y = parseFloat($destDiv.css('top'), 10);
-	_x = parseFloat($destDiv.css('left'), 10);
+	_y = parseInt($destDiv.css('top'), 10);
+	_x = parseInt($destDiv.css('left'), 10);
 	$destDiv.css('top', _y+layout_oy+'px');
 	$destDiv.css('left', _x+layout_ox+'px');
 	$destDiv.css('z-index', '998');
